Break out LoungeMute to new ModerationController
